Fixed interface lock problem after purchasing an item. Made power-ups menu larger.

+++ MKStoreKit.m
@@ -175,7 +175,7 @@ static NSDictionary *errorDictionary;
     NSNumber *currentConsumableCount = self.purchaseRecord[consumableId];
     currentConsumableCount = @([currentConsumableCount doubleValue] - [creditCountToConsume doubleValue]);
     self.purchaseRecord[consumableId] = currentConsumableCount;
-	NSLog(@"New balance for consumable (%@): %ld", consumableId, currentConsumableCount.integerValue);
+	NSLog(@"New balance for consumable (%@): %ld", consumableId, (long)currentConsumableCount.integerValue);
     [self savePurchaseRecord];
     return currentConsumableCount;
 }
@@ -184,7 +184,7 @@ static NSDictionary *errorDictionary;
 	NSNumber *currentConsumableCount = self.purchaseRecord[consumableId];
 	currentConsumableCount = @([currentConsumableCount doubleValue] + [creditCountToAdd doubleValue]);
 	self.purchaseRecord[consumableId] = currentConsumableCount;
-	NSLog(@"New balance for consumable (%@): %ld", consumableId, currentConsumableCount.integerValue);
+	NSLog(@"New balance for consumable (%@): %ld", consumableId, (long)currentConsumableCount.integerValue);
 	[self savePurchaseRecord];
 	return currentConsumableCount;
 }
